Organised components & started home component
