+++ src/app/app.component.html
@@ -1,4 +1,4 @@
-<div class="h-screen relative dark:bg-slate-800">
+<div class="h-screen relative dark:bg-slate-800 transition-all duration-500 background-image">
 	<app-navbar class="absolute bottom-0 w-full"></app-navbar>
 	<router-outlet></router-outlet>
 </div>
