+++ src/app/components/home/home.component.html
@@ -1,0 +1,8 @@
+<div class="flex justify-center flex-col h-full items-center dark:text-slate-50 background-image bg-right bg-no-repeat bg-contain">
+	<h1 class="text-5xl text-center">Hi! I'm Adrián Rosón</h1>
+	<h2 class="text-3xl font-bold my-3">Junior Developer</h2>
+
+	<div class="flex">
+		<!-- <a href="https://github.com/AdrianR25" class="button">Github</a> -->
+	</div>
+</div>
