+++ src/app/components/home/home.component.scss
@@ -1,0 +1,5 @@
+.background-image {
+	background-image: url("../../../assets/Illustration.png");
+	//background-image: url("https://www.w3schools.com/images/colorpicker2000.png");
+	//background-color: red;
+}
