--- src/app/home/home.component.html
@@ -1,1 +1,0 @@
-<p>home works!</p>
